+++ app/controllers/reports_controller.rb
@@ -6,6 +6,7 @@ class ReportsController < HtmlController
     @available_organizations = policy_scope Organization.where(deleted_at: nil).order(:organization_name)
     @available_chapters = policy_scope Chapter.where(deleted_at: nil).order(:chapter_name)
     @available_groups = policy_scope Group.where(deleted_at: nil).order(:group_name)
+    @selected_student_id = params[:student_id]
 
     authorize Group, :index?
   end