+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 - Added ability to edit skill and grade descriptions
+- Added ability to generate student-level reports
 - Defaulted starting date for analytics to the last lesson's month
 - Replaced existing charts using the `Chart.js` library 
 - Fixed issue where group reports crashed when students had no summaries